# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AcctMarket - Social Media Account Selling Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USERS
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all active categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Category deleted

  # PRODUCTS
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: User orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Purchase a product (deducts from balance)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Insufficient balance or product unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details including delivered logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # WALLET
...
  /deposits/initiate:
    post:
      operationId: initiateDeposit
      tags: [deposits]
      summary: Initiate a Paystack deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateDepositBody"
      responses:
        "200":
          description: Paystack payment URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositInitResponse"
  /deposits/verify/{reference}:
    get:
      operationId: verifyDeposit
      tags: [deposits]
      summary: Verify Paystack payment and credit balance
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment verified and balance credited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositVerifyResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
      required:
        - username
        - email
        - password
        - fullName
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        balance:
          type: number
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - fullName
        - balance
        - isAdmin
        - isBanned
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        isActive:
          type: boolean
        productCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - isActive
        - productCount
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - slug
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        isActive:
          type: boolean
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productPrice:
          type: number
        status:
          type: string
        deliveredLogs:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - productId
        - productName
        - productPrice
        - status
        - createdAt
        - updatedAt
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - totalPages
    CreateOrderBody:
      type: object
      properties:
        productId:
          type: integer
      required:
        - productId
...
    InitiateDepositBody:
      type: object
      properties:
        amount:
          type: number
        callbackUrl:
          type: string
      required:
        - amount
    DepositInitResponse:
      type: object
      properties:
        authorizationUrl:
          type: string
        reference:
          type: string
        accessCode:
          type: string
      required:
        - authorizationUrl
        - reference
        - accessCode
    DepositVerifyResponse:
      type: object
      properties:
        status:
          type: string
        amount:
          type: number
        newBalance:
          type: number
        message:
          type: string
      required:
        - status
        - amount
        - newBalance
        - message
...