# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AcctMarket - Social Media Account Selling Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  productId: number;
}