# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AcctMarket - Social Media Account Selling Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  fullName: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  balance: number;
  isAdmin: boolean;
  isBanned: boolean;
  createdAt: string;
}